chore: refactor before sending PR
Changed the generic naming to ARS548 since it looked like many things can not be generalized
Added documentation to the new methods
Addressed remaining TODOs
