+++ nebula_decoders/src/nebula_decoders_continental/decoders/continental_ars548_decoder.cpp
@@ -14,7 +14,8 @@
 
 #include "nebula_decoders/nebula_decoders_continental/decoders/continental_ars548_decoder.hpp"
 
-#include "nebula_decoders/nebula_decoders_continental/decoders/continental_ars548.hpp"
+#include "nebula_common/continental/continental_ars548.hpp"
+#include "nebula_common/continental/continental_common.hpp"
 
 #include <cmath>
 #include <utility>
@@ -26,8 +27,7 @@ namespace drivers
 namespace continental_ars548
 {
 ContinentalARS548Decoder::ContinentalARS548Decoder(
-  const std::shared_ptr<drivers::ContinentalRadarEthernetSensorConfiguration> &
-    sensor_configuration)
+  const std::shared_ptr<drivers::ContinentalARS548SensorConfiguration> & sensor_configuration)
 {
   sensor_configuration_ = sensor_configuration;
 }
