+++ nebula_ros/CMakeLists.txt
@@ -126,7 +126,7 @@ rclcpp_components_register_node(robosense_hw_monitor_ros_wrapper
 ## Continental
 # Hw Interface
 ament_auto_add_library(continental_hw_ros_wrapper SHARED
-        src/continental/continental_radar_ethernet_hw_interface_ros_wrapper.cpp
+        src/continental/continental_ars548_hw_interface_ros_wrapper.cpp
         )
 rclcpp_components_register_node(continental_hw_ros_wrapper
         PLUGIN "ContinentalRadarEthernetHwInterfaceRosWrapper"
@@ -135,7 +135,7 @@ rclcpp_components_register_node(continental_hw_ros_wrapper
 
 # DriverDecoder
 ament_auto_add_library(continental_driver_ros_wrapper SHARED
-        src/continental/continental_radar_ethernet_decoder_ros_wrapper.cpp
+        src/continental/continental_ars548_decoder_ros_wrapper.cpp
         )
 rclcpp_components_register_node(continental_driver_ros_wrapper
         PLUGIN "ContinentalRadarEthernetDriverRosWrapper"
