+++ nebula_ros/src/robosense/robosense_hw_monitor_ros_wrapper.cpp
@@ -52,7 +52,7 @@ Status RobosenseHwMonitorRosWrapper::Shutdown()
 {
   return Status::OK;
 }
-//
+
 Status RobosenseHwMonitorRosWrapper::InitializeHwMonitor(
   const drivers::SensorConfigurationBase & sensor_configuration)
 {
